pyproject.toml
[project]
name = "multi-agent-rag"
version = "0.1.0"
description = "Multi-Agent RAG System with FastMCP Integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno[openai]>=2.0",
    "fastapi[standard]>=0.115",
    "uvicorn>=0.32",
    "fastmcp>=2.0",
    "pymilvus>=2.4",
    "pydantic-settings>=2.0",
    "pyjwt>=2.8",
    "openai>=1.0",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[project.scripts]
serve = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
